requirements.txt
-i https://pypi.org/simple
certifi==2018.10.15
chardet==3.0.4
idna==2.7
python-dotenv==0.9.1
requests==2.20.0
urllib3==1.24
